[] Separated classes into different modules.
#3
